+++ ProjectEuler/Problem1.h
@@ -4,7 +4,7 @@
 
 class PROJECTEULER_API Problem1 {
 public:
-    static unsigned long SumMultiplesOf3And5Below(unsigned long);
-    static unsigned long Sum(unsigned long);
-    static unsigned long SumDivisibleBy(unsigned long, unsigned long);
+    static unsigned long long SumMultiplesOf3And5Below(unsigned long long);
+    static unsigned long long Sum(unsigned long long);
+    static unsigned long long SumDivisibleBy(unsigned long long, unsigned long long);
 };
